Moves interfaces to types/interfaces.ts
